+++ docs/changelog.rst
@@ -28,6 +28,7 @@ New features:
   ``mb_releasegroupid`` as well as simulates track ids using release id
   and tracklist positions. Track ids are stored in ``mb_trackid``. :bug:`#2336`
   Thanks to :user:`dbogdanov`.
+* :doc:`/plugins/discogs`: Fetch original year from master releases. :bug:`#1122`
 * As a first step to get :bug:`#406` implemented, beets now imports the
   ``musicbrainz_releasetrackid`` field into the library and tags media files
   accordingly. Thanks to :user:`Rawrmonkeys`.